# backend_go/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.BuyPayload:
    properties:
      external_product_id:
        type: string
      product_id:
        type: integer
      provider:
        type: string
      quantity:
        type: integer
      referral_bot_token:
        type: string
      user_id:
        type: integer
    required:
    - quantity
    - user_id
    type: object
  handlers.categoryPayload:
    properties:
      name:
        type: string
      parent_id:
        type: integer
    required:
    - name
    type: object
...
  handlers.depositRequest:
    properties:
      amount:
        type: number
      user_id:
        type: integer
    required:
    - amount
    - user_id
    type: object
...
  models.CategoryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      sub_categories:
        items:
          $ref: '#/definitions/models.CategoryResponse'
        type: array
    type: object
...
  models.OrderResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
      status:
        type: string
      user_id:
        type: integer
      user_telegram_id:
        type: integer
    type: object
  models.OrderSlimResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      status:
        type: string
      user_id:
        type: integer
    type: object
...
  responses.ErrorResponseSchema:
    description: This is the standard error response structure.
    properties:
      data: {}
      error:
        type: string
      success:
        example: false
        type: boolean
    type: object
  responses.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  responses.ResponseSchema-array_models_CategoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.CategoryResponse'
        type: array
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  responses.ResponseSchema-array_models_OrderResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.OrderResponse'
        type: array
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  responses.ResponseSchema-models_CategoryResponse:
    properties:
      data:
        $ref: '#/definitions/models.CategoryResponse'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  responses.ResponseSchema-responses_MessageResponse:
    properties:
      data:
        $ref: '#/definitions/responses.MessageResponse'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  responses.ResponseSchema-services_BuyResponse:
    properties:
      data:
        $ref: '#/definitions/services.BuyResponse'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  services.BuyResponse:
    properties:
      balance:
        type: number
      order:
        $ref: '#/definitions/models.OrderSlimResponse'
      product_name:
        type: string
      product_price:
        type: number
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for your project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Your Project API
  version: "1.0"
paths:
...
  /balance/deposit:
    post:
      consumes:
      - application/json
      description: Manually adds a deposit transaction to a user's balance. Prefer
        using the /deposit/invoice flow.
      parameters:
      - description: Deposit data
        in: body
        name: deposit
        required: true
        schema:
          $ref: '#/definitions/handlers.depositRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-responses_MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ServiceApiKeyAuth: []
      summary: (DEPRECATED) Deposit Balance
      tags:
      - Balance
  /categories:
    get:
      description: Get a hierarchical list of all categories.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-array_models_CategoryResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      - ServiceApiKeyAuth: []
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new category with the given name and optional parent_id.
      parameters:
      - description: Category Name and Parent ID
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/handlers.categoryPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.ResponseSchema-models_CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Create a new category
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Delete a category by its ID.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Delete a category
      tags:
      - Categories
    get:
      description: Get a single category by its ID.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-models_CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Get a category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update a category's name and/or parent by its ID.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: New Category Name and/or Parent ID
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/handlers.categoryPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-models_CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Update a category
      tags:
      - Categories
...
  /orders:
    get:
      description: Retrieves a list of all orders.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-array_models_OrderResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: List Orders
      tags:
      - Orders
  /orders/{id}/cancel:
    post:
      description: Cancels an order and refunds the user.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-responses_MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Cancel an Order
      tags:
      - Orders
  /orders/buy-from-balance:
    post:
      consumes:
      - application/json
      description: Creates an order and processes the purchase of a product from the
        user's balance.
      parameters:
      - description: Purchase data
        in: body
        name: purchase
        required: true
        schema:
          $ref: '#/definitions/handlers.BuyPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-services_BuyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ServiceApiKeyAuth: []
      summary: Buy a Product
      tags:
      - Orders
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ServiceApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"